refactor(step4): refactor input refine logic by responsibility

+++ src/main/kotlin/racingcar/view/RacingcarGameInputView.kt
@@ -12,7 +12,9 @@ object RacingcarGameInputView {
             .map { RacingcarGameInputValidator.validateCarName(it) }
 
         val roundNumberString = getInput("시도할 횟수는 몇 회인가요?")
-        val roundNumber = RacingcarGameInputValidator.validateRoundNumberString(roundNumberString)
+        val roundNumber = RacingcarGameInputValidator
+            .validateRoundNumberString(roundNumberString)
+            .toInt()
 
         return RacingcarGameInput(
             carNames = carNames,
